--- a/aten/src/ATen/native/cuda/CompareEQKernels.cu
+++ b/aten/src/ATen/native/cuda/CompareEQKernels.cu
@@ -29,6 +29,13 @@ struct CompareFunctor{
   }
 };
 
+// template<typename scalar_t>
+// struct CompareNEFunctor {
+//   __device__ __forceinline__ bool operator() (scalar_t a, scalar_t b) const {
+//     return a != b;
+//   }
+// };
+
 void ne_kernel_cuda(TensorIteratorBase& iter) {
   AT_DISPATCH_ALL_TYPES_AND_COMPLEX_AND3(kHalf, kBFloat16, kBool, iter.common_dtype(), "ne_cuda", [&]() {
     gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(CompareEqOpType::NE));