--- a/aten/src/ATen/native/cuda/CompareEQKernels.cu
+++ b/aten/src/ATen/native/cuda/CompareEQKernels.cu
@@ -11,19 +11,14 @@
 
 namespace at { namespace native {
 
-namespace {
-  enum class CompareEqOpType {NE, EQ};
-}
-
-
 template<typename scalar_t>
 struct CompareFunctor{
-  CompareFunctor(const CompareEqOpType op): op_(op) {}
-  const CompareEqOpType op_;
+  CompareFunctor(const int op): op_(op) {TORCH_INTERNAL_ASSERT_DEBUG_ONLY(op_>=0 && op_<=1);}
+  const int op_;
   __device__ __forceinline__ bool operator() (scalar_t a, scalar_t b) const {
-    if (op_ == CompareEqOpType::NE) {
+    if (op_ == 0) {
       return a != b;
-    } else { //EQ
+    } else if (op_ == 1) { //EQ
       return a == b;
     }
   }
@@ -31,13 +26,13 @@ struct CompareFunctor{
 
 void ne_kernel_cuda(TensorIteratorBase& iter) {
   AT_DISPATCH_ALL_TYPES_AND_COMPLEX_AND3(kHalf, kBFloat16, kBool, iter.common_dtype(), "ne_cuda", [&]() {
-    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(CompareEqOpType::NE));
+    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(0));
   });
 }
 
 void eq_kernel_cuda(TensorIteratorBase& iter) {
   AT_DISPATCH_ALL_TYPES_AND_COMPLEX_AND3(kHalf, kBFloat16, kBool, iter.common_dtype(), "ne_cuda", [&]() {
-    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(CompareEqOpType::EQ));
+    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(1));
   });
 }