This unifies 6 compare ops (NE, EQ, LT, LE, GE, GT) into 2 kernels, reducing context size. Performance is better for broadcasted cases, on-par for non-broadcasted
With this PR, benchmarks for contiguous, 1M-MM, 1M-M1, op with scalar (size in MB and bandwidth in GB/s):
```
(7, 0)
    5.0,   518.4
   10.0,   666.1
   15.0,   704.4
   20.0,   734.3
   25.0,   746.5
   30.0,   757.4
   35.0,   762.6
   40.0,   770.6
   45.0,   775.0
   50.0,   780.3
   55.0,   783.1
   60.0,   785.5
   65.0,   787.9
   70.0,   788.9
   75.0,   790.6
   80.0,   790.9
   85.0,   794.2
   90.0,   795.9
   95.0,   797.1
  100.0,   798.7
    3.0,   403.8     1.0,   136.3     3.0,   388.5
    6.0,   464.2     2.0,   158.9     6.0,   771.6
    9.0,   473.9     3.0,   169.4     9.0,   691.1
   12.0,   505.2     4.0,   174.3    12.0,   687.5
   15.0,   507.2     5.0,   178.0    15.0,   723.7
   18.0,   521.1     6.0,   180.1    18.0,   736.6
   21.0,   527.8     7.0,   181.8    21.0,   744.8
   24.0,   538.6     8.0,   183.3    24.0,   757.0
   27.0,   541.0     9.0,   184.2    27.0,   762.5
   30.0,   539.1    10.0,   185.0    30.0,   770.3
   33.0,   543.9    11.0,   185.2    33.0,   775.7
   36.0,   544.9    12.0,   185.4    36.0,   781.3
   39.0,   547.4    13.0,   186.1    39.0,   785.0
   42.0,   544.1    14.0,   185.9    42.0,   788.4
   45.0,   546.8    15.0,   186.3    45.0,   790.3
   48.0,   554.3    16.0,   187.2    48.0,   793.4
   51.0,   552.5    17.0,   186.9    51.0,   795.4
   54.0,   549.4    18.0,   186.7    54.0,   797.8
   57.0,   549.2    19.0,   186.8    57.0,   799.6
   60.0,   552.2    20.0,   187.7    60.0,   801.8

```
Master
```
    5.0,   743.4
   10.0,   665.7
   15.0,   702.3
   20.0,   727.5
   25.0,   740.7
   30.0,   757.5
   35.0,   760.3
   40.0,   768.5
   45.0,   775.7
   50.0,   776.8
   55.0,   781.1
   60.0,   786.5
   65.0,   786.8
   70.0,   790.1
   75.0,   789.7
   80.0,   789.1
   85.0,   793.2
   90.0,   793.8
   95.0,   795.9
  100.0,   796.0
    3.0,   383.1     1.0,   129.0     3.0,   337.0
    6.0,   445.0     2.0,   149.6     6.0,   670.6
    9.0,   445.3     3.0,   159.6     9.0,   678.6
   12.0,   474.9     4.0,   164.1    12.0,   705.5
   15.0,   480.8     5.0,   167.2    15.0,   718.3
   18.0,   490.3     6.0,   169.1    18.0,   733.3
   21.0,   493.9     7.0,   168.5    21.0,   742.5
   24.0,   503.8     8.0,   171.9    24.0,   756.4
   27.0,   506.7     9.0,   171.3    27.0,   759.8
   30.0,   508.7    10.0,   172.4    30.0,   767.1
   33.0,   515.7    11.0,   174.2    33.0,   773.7
   36.0,   516.7    12.0,   170.4    36.0,   781.7
   39.0,   519.1    13.0,   174.4    39.0,   782.1
   42.0,   515.7    14.0,   174.1    42.0,   787.0
   45.0,   519.2    15.0,   172.7    45.0,   788.1
   48.0,   522.2    16.0,   175.4    48.0,   791.7
   51.0,   519.6    17.0,   175.1    51.0,   795.7
   54.0,   518.5    18.0,   174.8    54.0,   795.8
   57.0,   519.1    19.0,   174.4    57.0,   796.6
   60.0,   521.5    20.0,   175.6    60.0,   800.1
```
<details>
<summary>Benchmarking script </summary>

```
import torch
from matplotlib import pyplot as plt
from torch.utils.benchmark import Timer, Compare
import math
import click
print(torch.cuda.get_device_capability()) # check that we are on Volta (compute capability 7,0)
#torch.cuda.set_device(1)
# don't benchmark on anything too small, you'll see only overhead
@click.command()
@click.option('--op_str', default="torch.gt")
@click.option('--dtype_str', default="float", type=click.Choice(['float', 'half']))
def bench(op_str, dtype_str):
    if dtype_str == "float":
        dtype = torch.float
    elif dtype_str == "half":
        dtype = torch.half

    MB = 1024 * 1024
    size = MB
    results = []
    sizes = []
    for _ in range(20):
        torch.cuda.memory.empty_cache()
        a=torch.randn(int(size), device="cuda", dtype=dtype)
        b=torch.randn(int(size), device="cuda", dtype=dtype)
        t = Timer(stmt=f"{op_str}(a,b)", label = op_str, sub_label=f"{size/MB} MB", description="contiguous", globals = {"a":a, "b":b})
        res = t.blocked_autorange()
        results.append(res)
        sizes.append(size)
        size +=  MB
        del a #to save memory for next iterations
        del b
    c=Compare(results)
    #print(c)
    bw=[]
    bytes=[]
    element_size = torch.tensor([], dtype=dtype).element_size()
    output_element_size = 1
    for res, size in zip(results,sizes):
        bytes_io = 2*size*element_size + output_element_size * size
        bytes.append(bytes_io/MB)
        # we'll report bandwidth in GB/s
        bw.append(bytes_io/res.median * 1e-9)
        print(f"{bytes_io/MB:7.1f}, {bw[-1]:7.1f}")

    sizes = []
    results = [[],[],[]]

    size = MB
    for _ in range(20):
        torch.cuda.memory.empty_cache()
        M = math.floor(math.sqrt(size))
        a=torch.randn(1, M, device="cuda", dtype=dtype)
        b=torch.randn(M, M, device="cuda", dtype=dtype)
        b1 = torch.randn(M, 1, device="cuda", dtype=dtype)
        tb = Timer(stmt=f"{op_str}(a,b)", label = op_str, sub_label=f"{M*M/MB} MB", description="MMM1", globals = {"a":a, "b":b})
        t1 = Timer(stmt=f"{op_str}(a,b1)", label = op_str, sub_label=f"{M*M/MB} MB", description="M11M", globals = {"a":a, "b1":b1})
        ts = Timer(stmt=f"{op_str}(b,1.)", label = op_str, sub_label=f"{M*M/MB} MB", description="scalar", globals = {"a":a, "b":b})

        res = [t.blocked_autorange() for t in (tb, t1, ts)]
        for (rl, r) in zip(results, res):
            rl.append(r)
        sizes.append(M)
        size += MB
        del a #to save memory for next iterations
        del b
    comps = [Compare(r) for r in results]
    #[print(c) for c in comps]
    bw=[[],[],[]]

    for res, res1, ress, size in zip(results[0],results[1],results[2], sizes):
        bytes_io = (size+size*size)*element_size + output_element_size * size*size #(size+size+size*size)*4
        bytes_io1 = (size+size)*element_size + output_element_size * size*size #(size+size+size*size)*4
        bytes_ios = (size*size)*element_size + output_element_size * size * size
        bytes_iol = (bytes_io, bytes_io1, bytes_ios)
        for (bw_elem, bytes_elem, res_elem) in zip(bw, bytes_iol, (res, res1, ress)):
            bw_elem.append(bytes_elem/res_elem.median * 1e-9)
        print(f"{bytes_iol[0]/MB:7.1f}, {bw[0][-1]:7.1f}", f"{bytes_iol[1]/MB:7.1f}, {bw[1][-1]:7.1f}",
        f"{bytes_iol[2]/MB:7.1f}, {bw[2][-1]:7.1f}")

if __name__ == '__main__':
    bench()
```
</details>